Removed redundancy in text generation script
Reworked the get_frequencies function in text_stats.py, so that it could be called by the generate_text.py script to get the frequency for a single word rather than all words in the text. This replaced a redundancy in generate_text.py where these calculations where also implemented.